Miscellaneous cleanup for the Android codebase: - update gradle plugins versions - add formatting rules for AndroidManifest and gradle build files - cleanup java_godot_lib_jni
Note: logic was mostly moved around and no new logic/functionality was added.
