+++ platform/android/SCsub
@@ -18,6 +18,8 @@ android_files = [
     'java_class_wrapper.cpp',
     'java_godot_wrapper.cpp',
     'java_godot_io_wrapper.cpp',
+    'jni_utils.cpp',
+    'android_keys_utils.cpp'
 ]
 
 env_android = env.Clone()
