+++ platform/android/java/app/config.gradle
@@ -1,24 +1,24 @@
 ext.versions = [
-        androidGradlePlugin : '3.4.2',
-        compileSdk : 28,
-        minSdk : 18,
-        targetSdk : 28,
-        buildTools : '28.0.3',
-        supportCoreUtils : '28.0.0'
+    androidGradlePlugin: '3.6.0',
+    compileSdk         : 29,
+    minSdk             : 18,
+    targetSdk          : 29,
+    buildTools         : '29.0.1',
+    supportCoreUtils   : '28.0.0'
 
 ]
 
 ext.libraries = [
-        androidGradlePlugin : "com.android.tools.build:gradle:$versions.androidGradlePlugin",
-        supportCoreUtils : "com.android.support:support-core-utils:$versions.supportCoreUtils"
+    androidGradlePlugin: "com.android.tools.build:gradle:$versions.androidGradlePlugin",
+    supportCoreUtils   : "com.android.support:support-core-utils:$versions.supportCoreUtils"
 ]
 
 ext.getExportPackageName = { ->
-	// Retrieve the app id from the project property set by the Godot build command.
-	String appId = project.hasProperty("export_package_name") ? project.property("export_package_name") : ""
-	// Check if the app id is valid, otherwise use the default.
-	if (appId == null || appId.isEmpty()) {
-		appId = "com.godot.game"
-	}
-	return appId
+    // Retrieve the app id from the project property set by the Godot build command.
+    String appId = project.hasProperty("export_package_name") ? project.property("export_package_name") : ""
+    // Check if the app id is valid, otherwise use the default.
+    if (appId == null || appId.isEmpty()) {
+        appId = "com.godot.game"
+    }
+    return appId
 }
